fix: repository interface 분리 custom repository interface 분리
